themes versioning, navigation to details fragment with no shared viewmodel TODO: safeargs - done details screen remake Work, Service, Notifications
