+++ .idea/misc.xml
@@ -38,13 +38,13 @@
         <entry key="..\:/Users/marbleVHS/AndroidStudioProjects/ClairSavedImages/app/src/main/res/layout-horizontal/image_details_fragment.xml" value="0.22780797101449277" />
         <entry key="..\:/Users/marbleVHS/AndroidStudioProjects/ClairSavedImages/app/src/main/res/layout-land/image_details_fragment.xml" value="0.24382716049382716" />
         <entry key="..\:/Users/marbleVHS/AndroidStudioProjects/ClairSavedImages/app/src/main/res/layout-land/login_fragment.xml" value="0.29717813051146386" />
-        <entry key="..\:/Users/marbleVHS/AndroidStudioProjects/ClairSavedImages/app/src/main/res/layout/favourites_list_fragment.xml" value="0.33" />
+        <entry key="..\:/Users/marbleVHS/AndroidStudioProjects/ClairSavedImages/app/src/main/res/layout/favourites_list_fragment.xml" value="0.21" />
         <entry key="..\:/Users/marbleVHS/AndroidStudioProjects/ClairSavedImages/app/src/main/res/layout/fragment_image_details.xml" value="0.19270833333333334" />
         <entry key="..\:/Users/marbleVHS/AndroidStudioProjects/ClairSavedImages/app/src/main/res/layout/fragment_login.xml" value="0.33" />
         <entry key="..\:/Users/marbleVHS/AndroidStudioProjects/ClairSavedImages/app/src/main/res/layout/fragment_nav_bar.xml" value="0.2667572463768116" />
         <entry key="..\:/Users/marbleVHS/AndroidStudioProjects/ClairSavedImages/app/src/main/res/layout/fragment_profile.xml" value="0.288034188034188" />
         <entry key="..\:/Users/marbleVHS/AndroidStudioProjects/ClairSavedImages/app/src/main/res/layout/image_card_view.xml" value="0.25" />
-        <entry key="..\:/Users/marbleVHS/AndroidStudioProjects/ClairSavedImages/app/src/main/res/layout/image_details_fragment.xml" value="0.25" />
+        <entry key="..\:/Users/marbleVHS/AndroidStudioProjects/ClairSavedImages/app/src/main/res/layout/image_details_fragment.xml" value="0.18" />
         <entry key="..\:/Users/marbleVHS/AndroidStudioProjects/ClairSavedImages/app/src/main/res/layout/image_list_fragment.xml" value="0.22" />
         <entry key="..\:/Users/marbleVHS/AndroidStudioProjects/ClairSavedImages/app/src/main/res/layout/image_view.xml" value="0.22" />
         <entry key="..\:/Users/marbleVHS/AndroidStudioProjects/ClairSavedImages/app/src/main/res/layout/item_error.xml" value="0.24048913043478262" />
