+++ app/build.gradle
@@ -2,6 +2,7 @@ plugins {
     id 'com.android.application'
     id 'kotlin-android'
     id 'kotlin-kapt'
+    id 'kotlin-parcelize'
     id 'androidx.navigation.safeargs'
 }
 
@@ -47,59 +48,64 @@ android {
 
 dependencies {
 
-    implementation 'com.github.kirich1409:viewbindingpropertydelegate-noreflection:1.5.6'
-
-    implementation 'androidx.appcompat:appcompat:1.4.1'
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
-    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.4.1'
-    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.4.1'
-    implementation 'androidx.core:core-ktx:1.7.0'
-    implementation 'androidx.work:work-runtime-ktx:2.7.1'
-
-    implementation 'androidx.activity:activity-ktx:1.4.0'
-    implementation 'androidx.fragment:fragment-ktx:1.4.1'
-
-    implementation 'com.google.android.material:material:1.6.0-rc01'
-
-    implementation 'androidx.swiperefreshlayout:swiperefreshlayout:1.1.0'
-
-    implementation 'androidx.navigation:navigation-fragment-ktx:2.4.2'
-    implementation 'androidx.navigation:navigation-ui-ktx:2.4.2'
-
-    implementation 'com.google.dagger:dagger:2.41'
-    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    kapt 'com.google.dagger:dagger-compiler:2.41'
 
 
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    implementation(project(":core:ui"))
+    
 
-    implementation 'com.squareup.okhttp3:okhttp:5.0.0-alpha.6'
-    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
-    implementation 'com.squareup.okhttp3:logging-interceptor:5.0.0-alpha.6'
-    implementation 'io.coil-kt:coil:1.1.0'
-    implementation 'io.coil-kt:coil-gif:1.4.0'
-    implementation 'com.google.code.gson:gson:2.9.0'
-    implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
-    implementation 'androidx.paging:paging-runtime-ktx:3.1.1'
+    implementation ("com.github.kirich1409:viewbindingpropertydelegate-noreflection:1.5.6")
 
-    implementation 'androidx.datastore:datastore-preferences:1.0.0'
+    implementation ("androidx.appcompat:appcompat:1.4.1")
+    implementation ("androidx.constraintlayout:constraintlayout:2.1.3")
+    implementation ("androidx.lifecycle:lifecycle-viewmodel-ktx:2.4.1")
+    implementation ("androidx.lifecycle:lifecycle-runtime-ktx:2.4.1")
+    implementation ("androidx.core:core-ktx:1.7.0")
+    implementation ("androidx.work:work-runtime-ktx:2.7.1")
 
-    implementation 'androidx.room:room-runtime:2.4.2'
-    kapt 'androidx.room:room-compiler:2.4.2'
+    implementation ("androidx.activity:activity-ktx:1.4.0")
+    implementation ("androidx.fragment:fragment-ktx:1.4.1")
 
-    implementation 'androidx.room:room-ktx:2.4.2'
-    implementation 'androidx.room:room-paging:2.4.2'
+    implementation ("com.google.android.material:material:1.7.0-alpha01")
 
-    implementation 'org.jetbrains.kotlin:kotlin-stdlib:1.6.21'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.1-native-mt'
+    implementation ("androidx.swiperefreshlayout:swiperefreshlayout:1.1.0")
+
+    implementation ("androidx.navigation:navigation-fragment-ktx:2.4.2")
+    implementation ("androidx.navigation:navigation-ui-ktx:2.4.2")
+
+    implementation ("com.google.dagger:dagger:2.41")
+    implementation ("androidx.legacy:legacy-support-v4:1.0.0")
+    kapt ("com.google.dagger:dagger-compiler:2.41")
 
 
-    implementation 'com.github.MikeOrtiz:TouchImageView:3.1.0'
+    testImplementation ("junit:junit:4.13.2")
+    androidTestImplementation ("androidx.test.ext:junit:1.1.3")
+    androidTestImplementation ("androidx.test.espresso:espresso-core:3.4.0")
 
-    implementation 'com.vk:android-sdk-core:3.5.2'
-    implementation 'com.vk:android-sdk-api:3.5.2'
+    implementation ("com.squareup.okhttp3:okhttp:5.0.0-alpha.7")
+    implementation ("com.squareup.retrofit2:retrofit:2.9.0")
+    implementation ("com.squareup.okhttp3:logging-interceptor:5.0.0-alpha.7")
+    implementation ("io.coil-kt:coil:1.1.0")
+    implementation ("io.coil-kt:coil-gif:1.4.0")
+    implementation ("com.google.code.gson:gson:2.9.0")
+    implementation ("com.squareup.retrofit2:converter-gson:2.9.0")
+    implementation ("androidx.paging:paging-runtime-ktx:3.1.1")
+
+    implementation ("androidx.datastore:datastore-preferences:1.0.0")
+
+    implementation ("androidx.room:room-runtime:2.4.2")
+    kapt ("androidx.room:room-compiler:2.4.2")
+
+    implementation ("androidx.room:room-ktx:2.4.2")
+    implementation ("androidx.room:room-paging:2.4.2")
+
+    implementation ("org.jetbrains.kotlin:kotlin-stdlib:1.6.21")
+    implementation ("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.1-native-mt")
+
+
+    implementation ("com.github.MikeOrtiz:TouchImageView:3.1.0")
+
+    implementation ("com.vk:android-sdk-core:3.5.2")
+    implementation ("com.vk:android-sdk-api:3.5.2")
 
 }
