+++ app/src/main/res/layout/image_details_fragment.xml
@@ -12,6 +12,7 @@
         android:id="@+id/ivSelectedImage"
         android:layout_width="0dp"
         android:layout_height="0dp"
+        android:transitionName="chosen_image"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
@@ -28,8 +29,8 @@
         android:layout_marginBottom="32dp"
         android:button="@drawable/like_button_drawable"
         android:fitsSystemWindows="true"
-        android:minWidth="0dp"
-        android:minHeight="0dp"
+        android:minWidth="48dp"
+        android:minHeight="48dp"
         android:textIsSelectable="false"
         android:visibility="invisible"
         app:layout_constraintBottom_toBottomOf="parent"
@@ -43,8 +44,8 @@
         android:layout_height="wrap_content"
         android:button="@drawable/favourites_button_drawable"
         android:fitsSystemWindows="true"
-        android:minWidth="0dp"
-        android:minHeight="0dp"
+        android:minWidth="48dp"
+        android:minHeight="48dp"
         android:visibility="invisible"
         app:layout_constraintBottom_toBottomOf="@+id/zoomInButton"
         app:layout_constraintEnd_toEndOf="@+id/ivSelectedImage"
