+++ core/ui/src/main/res/values-night/themes.xml
@@ -1,5 +1,6 @@
 <resources>
-    <style name="Theme.ClairSavedImages" parent="Theme.Material3.Dark.NoActionBar">
+
+    <style name="Base.V23.Night.Theme.ClairSavedImages" parent="Theme.Material3.Dark.NoActionBar">
         <item name="android:windowLightStatusBar">false</item>
         <item name="android:statusBarColor">@color/transparent</item>
         <item name="android:navigationBarColor">@color/transparent</item>
@@ -30,4 +31,7 @@
         <item name="colorSurfaceInverse">@color/md_theme_dark_inverseSurface</item>
         <item name="colorPrimaryInverse">@color/md_theme_dark_primaryInverse</item>
     </style>
+
+    <style name="Base.Theme.ClairSavedImages" parent="Base.V23.Night.Theme.ClairSavedImages"/>
+
 </resources>
