+++ core/ui/src/main/res/values/themes.xml
@@ -1,5 +1,6 @@
 <resources>
-    <style name="Theme.ClairSavedImages" parent="Theme.Material3.Light.NoActionBar">
+
+    <style name="Base.V23.Theme.ClairSavedImages" parent="Theme.Material3.Light.NoActionBar">
         <item name="android:windowLightStatusBar">true</item>
         <item name="android:statusBarColor">@color/transparent</item>
         <item name="android:navigationBarColor">@color/transparent</item>
@@ -30,4 +31,9 @@
         <item name="colorSurfaceInverse">@color/md_theme_light_inverseSurface</item>
         <item name="colorPrimaryInverse">@color/md_theme_light_primaryInverse</item>
     </style>
+
+    <style name="Base.Theme.ClairSavedImages" parent="Base.V23.Theme.ClairSavedImages"/>
+
+    <style name="Theme.ClairSavedImages" parent="Base.Theme.ClairSavedImages"/>
+
 </resources>
