+++ settings.gradle
@@ -8,3 +8,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "ClairSavedImages"
 include ':app'
+include ':core:ui'
